* note: unibot can now be build in c++0x mode and with libc++  * added windows support  * added lua 5.2 support (modified scripts accordingly)  * modularized lua includes (global, exttable, game_framework)  * updated LUA_PATH setting  * added bot restart capability (unix only)  * fixed some threading/lock issues  * added list_scripts (prints all loaded lua scripts)  * added get_os() function for scripts  * include c++ stdlib in version info (for libstdc++ and libc++)  * updated clang info  * (c) year ++  * misc cleanup and fixes

+++ config.cpp
@@ -1,6 +1,6 @@
 /*
  *  UniBot
- *  Copyright (C) 2009 - 2010 Florian Ziesche
+ *  Copyright (C) 2009 - 2011 Florian Ziesche
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -141,7 +141,7 @@ string config::get_hostname() {
 }
 
 unsigned short int config::get_port() {
-	return strtoul(config_data["port"].c_str(), NULL, 10);
+	return (unsigned short int)strtoul(config_data["port"].c_str(), NULL, 10);
 }
 
 string config::get_channel() {
